+++ config.c
@@ -2158,11 +2158,7 @@ config_load(struct config *conf, const char *conf_path,
     bool ret = false;
 
     *conf = (struct config) {
-#ifdef HAVE_TERMINFO
-        .term = xstrdup("foot"),
-#else
-        .term = xstrdup("xterm-256color"),
-#endif
+    	.term = xstrdup(DEFAULT_TERM),
         .shell = get_shell(),
         .title = xstrdup("foot"),
         .app_id = xstrdup("foot"),