+++ main.c
@@ -61,7 +61,7 @@ print_usage(const char *prog_name)
         "  -c,--config=PATH                        load configuration from PATH ($XDG_CONFIG_HOME/foot/foot.ini)\n"
         "  -C,--check-config                       verify configuration, exit with 0 if ok, otherwise exit with 1\n"
         "  -f,--font=FONT                          comma separated list of fonts in fontconfig format (monospace)\n"
-        "  -t,--term=TERM                          value to set the environment variable TERM to (foot)\n"
+        "  -t,--term=TERM                          value to set the environment variable TERM to (%s)\n"
         "  -T,--title=TITLE                        initial window title (foot)\n"
         "  -a,--app-id=ID                          window application ID (foot)\n"
         "  -m,--maximized                          start in maximized mode\n"
@@ -78,7 +78,7 @@ print_usage(const char *prog_name)
         "  -l,--log-colorize=[{never|always|auto}] enable/disable colorization of log output on stderr\n"
         "  -s,--log-no-syslog                      disable syslog logging (only applicable in server mode)\n"
         "  -v,--version                            show the version number and quit\n",
-        prog_name, prog_name);
+        prog_name, prog_name, DEFAULT_TERM);
 }
 
 bool